Progress actualization: 1. The functionality to load a new game from a .txt file has been implemented. 2. The interfaz has been improved, corrections and images have been added. 3. Corrections about the model of the problem have been done. We still have problems with the managment of thread ans javafx.

+++ src/customsThread/PacManThread.java
@@ -10,7 +10,7 @@ public class PacManThread extends Thread{
 	private PacManController pacManController; 
 	private PacMan uiPacMan;
 	
-	public PacManThread(int i, PacManController pacManController2, PacManModel pMx, PacMan pCx) {
+	public PacManThread(long i, PacManController pacManController2, PacManModel pMx, PacMan pCx) {
 	
 		sleep = i;
 		pacMan = pMx;
